{"schema":"decodex/commit/1","summary":"Split MCP lane control modules","authority":"XY-1115"}
